Extract ObjectInstance
