+++ lib/toy/object.rb
@@ -1,4 +1,5 @@
 require "toy/behaviours"
+require "toy/object_instance"
 
 module Toy
   Object = BasicObject.new
@@ -33,25 +34,6 @@ module Toy
 
     def new
       ObjectInstance.new(self)
-    end
-  end
-
-  class ObjectInstance < BasicObject
-    # Object instance methods
-    include Behaviours::InstanceVariables
-
-    # kind_of?
-    include Behaviours::ClassRelationships
-
-    # to_s and #inspect
-    include Behaviours::Inspection
-
-    def initialize(klass)
-      @klass = klass
-    end
-
-    def class
-      @klass
     end
   end
 end
